requirements.txt
# base deps
protobuf>=4.21.6,<=4.25.4

# Watch Dog 
watchdog
nvidia-ml-py

# Newer models need this
peft>=0.10
accelerate>=0.30
xformers

# utils
requests>=2.28.0
safetensors>=0.4.0
psutil>=5.9.0
ruamel.yaml>=0.17.21
scipy>=1.9.3
tqdm>=4.64.1
opencv-python-headless>=4.5.5.64
fast-histogram>=0.11
numpy<=1.26.4
uvloop

redis>=4.0.0

fastapi==0.115.0
python-multipart
uvicorn==0.30.6
pydantic==2.8.2
pillow
transformers
diffusers
torch

uvicorn[standard]
websocket-client

#rknn-toolkit-lite2==2.3.2
